Add files via upload
Gradient Descent is a generic optimization algorithm capable of finding optimal solution to a wide range of problems. The general idea of Gradient Descent is to tweak parameters iteratively in order to minimize a cost function. Training a model means searching for a combination of model parameters that minimizes a cost function (over the
training set). It is a search in the model’s parameters space , the more parameters a model has the more dimension this space has and the harder the search is.